refactor: AcceptType, Header 패키지 변경
