+++ app/src/main/java/nextstep/jwp/http/AcceptType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 public class AcceptType {
     public static final String TEXT_CSS = "text/css";
